+++ src/mimic/orm/macros/src/lib.rs
@@ -59,7 +59,6 @@ macro_rules! macro_node {
 
 // macro macros
 macro_node!(canister, node::Canister);
-macro_node!(constant, node::Constant);
 macro_node!(entity, node::Entity);
 macro_node!(enum_, node::Enum);
 macro_node!(enum_hash, node::EnumHash);