+++ src/mimic/orm/schema/src/node/schema.rs
@@ -1,8 +1,8 @@
 use crate::{
     node::{
-        Canister, Constant, Def, Entity, Enum, EnumHash, EnumValue, Error, Fixture, MacroNode, Map,
-        Newtype, Permission, Primitive, Record, Role, Sanitizer, Store, Tuple, ValidateNode,
-        Validator, VisitableNode,
+        Canister, Def, Entity, Enum, EnumHash, EnumValue, Error, Fixture, MacroNode, Map, Newtype,
+        Permission, Primitive, Record, Role, Sanitizer, Store, Tuple, ValidateNode, Validator,
+        VisitableNode,
     },
     visit::Visitor,
 };
@@ -25,7 +25,6 @@ use types::{ErrorVec, Timestamp};
 #[derive(Clone, Debug, Serialize, Deserialize)]
 pub enum SchemaNode {
     Canister(Canister),
-    Constant(Constant),
     Entity(Entity),
     Enum(Enum),
     EnumHash(EnumHash),
@@ -47,7 +46,6 @@ impl SchemaNode {
     const fn def(&self) -> &Def {
         match self {
             Self::Canister(n) => &n.def,
-            Self::Constant(n) => &n.def,
             Self::Entity(n) => &n.def,
             Self::Enum(n) => &n.def,
             Self::EnumHash(n) => &n.def,
@@ -71,7 +69,6 @@ impl MacroNode for SchemaNode {
     fn as_any(&self) -> &dyn Any {
         match self {
             Self::Canister(n) => n.as_any(),
-            Self::Constant(n) => n.as_any(),
             Self::Entity(n) => n.as_any(),
             Self::Enum(n) => n.as_any(),
             Self::EnumHash(n) => n.as_any(),
@@ -97,7 +94,6 @@ impl VisitableNode for SchemaNode {
     fn drive<V: Visitor>(&self, v: &mut V) {
         match self {
             Self::Canister(n) => n.accept(v),
-            Self::Constant(n) => n.accept(v),
             Self::Entity(n) => n.accept(v),
             Self::Enum(n) => n.accept(v),
             Self::EnumHash(n) => n.accept(v),